fix DTYPE_MAP and torch_dtype to dtype for transformers 5.x

+++ probes/env.py
@@ -73,13 +73,13 @@ def load_model(model_id, dtype):
     tokenizer = AutoTokenizer.from_pretrained(model_id)
     try:
         model = AutoModelForCausalLM.from_pretrained(
-            model_id, torch_dtype=dtype, device_map="cuda",
+            model_id, dtype=dtype, device_map="cuda",
         )
     except (ValueError, KeyError) as e:
         print(f"AutoModelForCausalLM failed ({e}); trying multimodal loader.")
         from transformers import AutoModelForImageTextToText
         model = AutoModelForImageTextToText.from_pretrained(
-            model_id, torch_dtype=dtype, device_map="cuda",
+            model_id, dtype=dtype, device_map="cuda",
         )
     model.train(False)  # inference mode
     return tokenizer, model
